+++ base/dtypes2strings/docs/types/test.ts
@@ -16,30 +16,30 @@
 * limitations under the License.
 */
 
-import dtypes2enums = require( './index' );
+import dtypes2strings = require( './index' );
 
 
 // TESTS //
 
 // The function returns an array...
 {
-	dtypes2enums( [ 'float64', 'float64' ] ); // $ExpectType (number | null)[]
+	dtypes2strings( [ 'float64', 'float64' ] ); // $ExpectType (string | null)[]
 }
 
 // The compiler throws an error if the function is not provided an array-like object containing data types...
 {
-	dtypes2enums( '5' ); // $ExpectError
-	dtypes2enums( 5 ); // $ExpectError
-	dtypes2enums( true ); // $ExpectError
-	dtypes2enums( false ); // $ExpectError
-	dtypes2enums( null ); // $ExpectError
-	dtypes2enums( {} ); // $ExpectError
-	dtypes2enums( [ '5' ] ); // $ExpectError
-	dtypes2enums( ( x: number ): number => x ); // $ExpectError
+	dtypes2strings( '5' ); // $ExpectError
+	dtypes2strings( 5 ); // $ExpectError
+	dtypes2strings( true ); // $ExpectError
+	dtypes2strings( false ); // $ExpectError
+	dtypes2strings( null ); // $ExpectError
+	dtypes2strings( {} ); // $ExpectError
+	dtypes2strings( [ '5' ] ); // $ExpectError
+	dtypes2strings( ( x: number ): number => x ); // $ExpectError
 }
 
 // The compiler throws an error if the function is provided an unsupported number of arguments...
 {
-	dtypes2enums(); // $ExpectError
-	dtypes2enums( [ 'float64' ], {} ); // $ExpectError
+	dtypes2strings(); // $ExpectError
+	dtypes2strings( [ 'float64' ], {} ); // $ExpectError
 }