+++ base/dtypes2strings/lib/main.js
@@ -1,4 +1,4 @@
-/*
+/**
 * @license Apache-2.0
 *
 * Copyright (c) 2025 The Stdlib Authors.
@@ -16,33 +16,45 @@
 * limitations under the License.
 */
 
-// TypeScript Version: 4.1
+'use strict';
 
-/// <reference types="@stdlib/types"/>
+// MODULES //
 
-import { DataType } from '@stdlib/types/ndarray';
+var resolveStr = require( './../../../base/dtype-resolve-str' );
+
+
+// MAIN //
 
 /**
-* Resolves a list of data type enumeration constants.
+* Resolves a list of data type strings.
 *
 * ## Notes
 *
-* -   If the function is unable to resolve an enumeration constant for a provided data type, the corresponding element in the returned array will be `null`.
+* -   If the function is unable to resolve a data type string for a provided data type, the corresponding element in the returned array will be `null`.
 *
-* @param dtypes - list of data types
-* @returns results
+* @param {Array} dtypes - list of data types
+* @returns {Array} results
 *
 * @example
-* var out = dtypes2enums( [ 'float32', 'float64' ] );
+* var out = dtypes2strings( [ 'float32', 'float64' ] );
 * // returns [...]
 *
 * @example
-* var out = dtypes2enums( [ 'foo', 'bar' ] );
+* var out = dtypes2strings( [ 'foo', 'bar' ] );
 * // returns [ null, null ]
 */
-declare function dtypes2enums( dtypes: ArrayLike<DataType> ): Array<number|null>;
+function dtypes2strings( dtypes ) {
+	var out;
+	var i;
+
+	out = [];
+	for ( i = 0; i < dtypes.length; i++ ) {
+		out.push( resolveStr( dtypes[ i ] ) );
+	}
+	return out;
+}
 
 
 // EXPORTS //
 
-export = dtypes2enums;
+module.exports = dtypes2strings;